requirements.txt
requests>=2.31.0
PyQt5>=5.15.9
python-dateutil>=2.8.2
